Refactor heap sort
I misjudged what heaps and priorities queues are and how they related to
heap sort. Reading more, I'm getting at this point that a heap is just
an efficient implementation of a priority queue. And heap sort takes the
heap idea and reimplements it over an unordered list, doing an in-place
sort. Which is far more cool than what I had thought previously.